+++ .setup/cpp/cTMed-total-std-vec.cpp
@@ -6,20 +6,42 @@
 #include <RcppArmadillo.h>
 // [[Rcpp::depends(RcppArmadillo)]]
 // [[Rcpp::export(.TotalStdVec)]]
-arma::vec TotalStdVec(const arma::vec& v, const double& delta_t) {
+arma::vec TotalStdVec(const arma::vec& v, const double& delta_t,
+                      const bool& sigma_diag = false) {
   arma::uword q = v.n_elem;
-  arma::uword p = (-1 + std::sqrt(1 + 24 * q)) / 6;
+  arma::uword p = 0;
+  if (sigma_diag) {
+    double p_dbl = (-1.0 + std::sqrt(1.0 + 4.0 * q)) / 2.0;
+    p = static_cast<arma::uword>(std::round(p_dbl));
+    if (p * p + p != q) {
+      Rcpp::stop(
+          "When sigma_diag = TRUE, v must contain vec(phi) and diag(sigma).");
+    }
+  } else {
+    double p_dbl = (-1.0 + std::sqrt(1.0 + 24.0 * q)) / 6.0;
+    p = static_cast<arma::uword>(std::round(p_dbl));
+    if (p * p + p * (p + 1) / 2 != q) {
+      Rcpp::stop(
+          "When sigma_diag = FALSE, v must contain vec(phi) and vech(sigma).");
+    }
+  }
   arma::mat phi = arma::mat(v.subvec(0, p * p - 1)).reshape(p, p);
-  arma::vec sigma_vech = v.subvec(p * p, q - 1);
   arma::mat sigma(p, p, arma::fill::zeros);
-  arma::uword index = 0;
-  for (arma::uword j = 0; j < p; ++j) {
-    for (arma::uword i = j; i < p; ++i) {
-      sigma(i, j) = sigma_vech[index];
-      if (i != j) {
-        sigma(j, i) = sigma_vech[index];
+  if (sigma_diag) {
+    sigma.diag() = v.subvec(p * p, q - 1);
+  } else {
+    arma::vec sigma_vech = v.subvec(p * p, q - 1);
+    arma::uword index = 0;
+    for (arma::uword j = 0; j < p; ++j) {
+      for (arma::uword i = j; i < p; ++i) {
+        sigma(i, j) = sigma_vech[index];
+
+        if (i != j) {
+          sigma(j, i) = sigma_vech[index];
+        }
+
+        index++;
       }
-      index++;
     }
   }
   arma::mat total = arma::expmat(delta_t * phi);
@@ -27,8 +49,8 @@ arma::vec TotalStdVec(const arma::vec& v, const double& delta_t) {
   arma::sylvester(cov_eta, phi, phi.t(), sigma);
   arma::vec sqrt_diag = arma::sqrt(cov_eta.diag());
   arma::mat total_std = total;
-  for (size_t i = 0; i < total.n_rows; i++) {
-    for (size_t j = 0; j < total.n_cols; j++) {
+  for (arma::uword i = 0; i < total.n_rows; ++i) {
+    for (arma::uword j = 0; j < total.n_cols; ++j) {
       total_std(i, j) *= sqrt_diag(j) / sqrt_diag(i);
     }
   }